+++ src/mbgl/shaders/fill_outline.cpp
@@ -15,8 +15,8 @@ uniform vec2 u_world;
 varying vec2 v_pos;
 
 uniform lowp float a_outline_color_t;
-attribute lowp vec4 a_outline_color;
-varying lowp vec4 outline_color;
+attribute highp vec4 a_outline_color;
+varying highp vec4 outline_color;
 uniform lowp float a_opacity_t;
 attribute lowp vec2 a_opacity;
 varying lowp float opacity;
@@ -31,7 +31,7 @@ void main() {
 
 )MBGL_SHADER";
 const char* fill_outline::fragmentSource = R"MBGL_SHADER(
-varying lowp vec4 outline_color;
+varying highp vec4 outline_color;
 varying lowp float opacity;
 
 varying vec2 v_pos;
